// frontend/bindings/github.com/telesma-app/mds/model/models.ts
export enum AuthenticatorStatus {
    /**
     * The Go zero value for the underlying type of the enum.
     */
    $zero = "",

    AuthenticatorStatusNotFIDOCertified = "NOT_FIDO_CERTIFIED",
    AuthenticatorStatusFIDOCertified = "FIDO_CERTIFIED",
    AuthenticatorStatusUserVerificationBypass = "USER_VERIFICATION_BYPASS",
    AuthenticatorStatusAttestationKeyCompromise = "ATTESTATION_KEY_COMPROMISE",
    AuthenticatorStatusUserKeyRemoteCompromise = "USER_KEY_REMOTE_COMPROMISE",
    AuthenticatorStatusUserKeyPhysicalCompromise = "USER_KEY_PHYSICAL_COMPROMISE",
    AuthenticatorStatusUpdateAvailable = "UPDATE_AVAILABLE",
    AuthenticatorStatusRetired = "RETIRED",
    AuthenticatorStatusRevoked = "REVOKED",
    AuthenticatorStatusSelfAssertionSubmitted = "SELF_ASSERTION_SUBMITTED",
    AuthenticatorStatusFIDOCertifiedL1 = "FIDO_CERTIFIED_L1",
    AuthenticatorStatusFIDOCertifiedL1Plus = "FIDO_CERTIFIED_L1plus",
    AuthenticatorStatusFIDOCertifiedL2 = "FIDO_CERTIFIED_L2",
    AuthenticatorStatusFIDOCertifiedL2Plus = "FIDO_CERTIFIED_L2plus",
    AuthenticatorStatusFIDOCertifiedL3 = "FIDO_CERTIFIED_L3",
    AuthenticatorStatusFIDOCertifiedL3Plus = "FIDO_CERTIFIED_L3plus",
    AuthenticatorStatusFIPS140CertifiedL1 = "FIPS140_CERTIFIED_L1",
    AuthenticatorStatusFIPS140CertifiedL2 = "FIPS140_CERTIFIED_L2",
    AuthenticatorStatusFIPS140CertifiedL3 = "FIPS140_CERTIFIED_L3",
    AuthenticatorStatusFIPS140CertifiedL4 = "FIPS140_CERTIFIED_L4",
};

/**
 * BiometricStatusReport describes biometric certification status for an authenticator model.
 */
export class BiometricStatusReport {
    "certLevel": number;
    "modality": string;
    "effectiveDate"?: string | null;
    "certificationDescriptor"?: string | null;
    "certificateNumber"?: string | null;
    "certificationPolicyVersion"?: string | null;
    "certificationRequirementsVersion"?: string | null;

    /** Creates a new BiometricStatusReport instance. */
    constructor($$source: Partial<BiometricStatusReport> = {}) {
        if (!("certLevel" in $$source)) {
            this["certLevel"] = 0;
        }
        if (!("modality" in $$source)) {
            this["modality"] = "";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new BiometricStatusReport instance from a string or object.
     */
    static createFrom($$source: any = {}): BiometricStatusReport {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new BiometricStatusReport($$parsedSource as Partial<BiometricStatusReport>);
    }
}
// ...
export class LookupResult {
    "aaguid": uuid$0.UUID;
    "found": boolean;
    "entry"?: PayloadEntry | null;
    "blobNumber": number;
    "source": string;
    "cached": boolean;
    "cachedAt": string;

    /** Creates a new LookupResult instance. */
    constructor($$source: Partial<LookupResult> = {}) {
        if (!("aaguid" in $$source)) {
            this["aaguid"] = "";
        }
        if (!("found" in $$source)) {
            this["found"] = false;
        }
        if (!("blobNumber" in $$source)) {
            this["blobNumber"] = 0;
        }
        if (!("source" in $$source)) {
            this["source"] = "";
        }
        if (!("cached" in $$source)) {
            this["cached"] = false;
        }
        if (!("cachedAt" in $$source)) {
            this["cachedAt"] = "0001-01-01T00:00:00.000Z";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new LookupResult instance from a string or object.
     */
    static createFrom($$source: any = {}): LookupResult {
        const $$createField2_0 = $$createType5;
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        if ("entry" in $$parsedSource) {
            $$parsedSource["entry"] = $$createField2_0($$parsedSource["entry"]);
        }
        return new LookupResult($$parsedSource as Partial<LookupResult>);
    }
}
// ...
export class PayloadEntry {
    "aaid"?: string | null;
    "aaguid"?: uuid$0.UUID;
    "attestationCertificateKeyIdentifiers"?: string[];
    "metadataStatement"?: MetadataStatement;
    "biometricStatusReports"?: BiometricStatusReport[];
    "statusReports"?: StatusReport[];
    "timeOfLastStatusChange"?: string;
    "rogueListURL"?: string | null;
    "rogueListHash"?: string | null;

    /** Creates a new PayloadEntry instance. */
    constructor($$source: Partial<PayloadEntry> = {}) {

        Object.assign(this, $$source);
    }

    /**
     * Creates a new PayloadEntry instance from a string or object.
     */
    static createFrom($$source: any = {}): PayloadEntry {
        const $$createField2_0 = $$createType6;
        const $$createField3_0 = $$createType18;
        const $$createField4_0 = $$createType20;
        const $$createField5_0 = $$createType22;
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        if ("attestationCertificateKeyIdentifiers" in $$parsedSource) {
            $$parsedSource["attestationCertificateKeyIdentifiers"] = $$createField2_0($$parsedSource["attestationCertificateKeyIdentifiers"]);
        }
        if ("metadataStatement" in $$parsedSource) {
            $$parsedSource["metadataStatement"] = $$createField3_0($$parsedSource["metadataStatement"]);
        }
        if ("biometricStatusReports" in $$parsedSource) {
            $$parsedSource["biometricStatusReports"] = $$createField4_0($$parsedSource["biometricStatusReports"]);
        }
        if ("statusReports" in $$parsedSource) {
            $$parsedSource["statusReports"] = $$createField5_0($$parsedSource["statusReports"]);
        }
        return new PayloadEntry($$parsedSource as Partial<PayloadEntry>);
    }
}
// ...
export class StatusReport {
    "status": AuthenticatorStatus;
    "effectiveDate"?: string | null;
    "authenticatorVersion"?: number | null;
    "batchCertificate"?: string | null;
    "certificate"?: string | null;
    "url"?: string | null;
    "certificationDescriptor"?: string | null;
    "certificateNumber"?: string | null;
    "certificationPolicyVersion"?: string | null;
    "certificationProfiles"?: string[];
    "certificationRequirementsVersion"?: string | null;
    "sunsetDate"?: string | null;
    "fipsRevision"?: number | null;
    "fipsPhysicalSecurityLevel"?: number | null;

    /** Creates a new StatusReport instance. */
    constructor($$source: Partial<StatusReport> = {}) {
        if (!("status" in $$source)) {
            this["status"] = AuthenticatorStatus.$zero;
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new StatusReport instance from a string or object.
     */
    static createFrom($$source: any = {}): StatusReport {
        const $$createField9_0 = $$createType6;
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        if ("certificationProfiles" in $$parsedSource) {
            $$parsedSource["certificationProfiles"] = $$createField9_0($$parsedSource["certificationProfiles"]);
        }
        return new StatusReport($$parsedSource as Partial<StatusReport>);
    }
}
// ...
const $$createType4 = PayloadEntry.createFrom;
const $$createType5 = $Create.Nullable($$createType4);
const $$createType6 = $Create.Array($Create.Any);
// ...
const $$createType18 = MetadataStatement.createFrom;
const $$createType19 = BiometricStatusReport.createFrom;
const $$createType20 = $Create.Array($$createType19);
const $$createType21 = StatusReport.createFrom;
const $$createType22 = $Create.Array($$createType21);